Removed test menu command and moved some utilities into their own general uitilities folder
